requirements.txt
supabase
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
python-dotenv
pypdf
pymupdf
sentence-transformers
faiss-cpu
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pandas
python-multipart